phonetic_matcher/requirements.txt
fastapi
uvicorn
psycopg2-binary
cologne-phonetics
jellyfish
rapidfuzz
python-dotenv
python-multipart 